requirements.txt
pillow~=10.1.0
numpy~=1.26.2
scipy~=1.11.4
colormath~=3.0.0
scikit-image~=0.22.0